Add IntUtil and AccidentAvoidanceService

+++ src/main/java/com/iress/toyrobot/model/Robot.java
@@ -3,7 +3,7 @@ package com.iress.toyrobot.model;
 import java.util.List;
 
 public interface Robot {
-    void move(byte[] move);
+    void move(int[] move);
     void left();
     void right();
     List<String> report();
