+++ src/main/java/com/iress/toyrobot/model/TableTop.java
@@ -6,5 +6,5 @@ import lombok.Getter;
 @Getter
 @Builder
 public class TableTop {
-    private byte[] dimensions;
+    private int[] dimensions;
 }
